fix: change types
