+++ src/utils/types/global/global.d.ts
